Fixes Pirouette bug
Fixes Pirouette bug which was apparently caused by a poor design choice in the path.createFacingArray(). I told it to append the value of 5 at the end of the list as the final facing direction. This value should instead have been the same as the last move in the list. Apparently.